requirements.txt
requests
python-dotenv
google.generativeai
fastapi
uvicorn
pydantic